requirements.txt
python-telegram-bot==20.6
fastapi==0.110.2
uvicorn==0.29.0
openai==1.23.2
apscheduler==3.10.4
dateparser==1.2.0
